plugin-dotnet-agent/src/main/kotlin/jetbrains/buildServer/dotnet/MSBuildVSTestLoggerArgumentsProvider.kt
package jetbrains.buildServer.dotnet

import jetbrains.buildServer.agent.CommandLineArgument
import kotlin.coroutines.experimental.buildSequence

/**
 * Provides arguments to dotnet related to TeamCity logger.
 */

class MSBuildVSTestLoggerArgumentsProvider(
        private val _loggerResolver: LoggerResolver)
    : ArgumentsProvider {

    override val arguments: Sequence<CommandLineArgument>
        get() = buildSequence {
            _loggerResolver.resolve(ToolType.VSTest).parentFile?.let {
                yield(CommandLineArgument("/p:VSTestLogger=logger://teamcity"))
                yield(CommandLineArgument("/p:VSTestTestAdapterPath=."))
            }
        }
}